+++ scripts/lib/constants.js
@@ -216,6 +216,7 @@ const skillOptions = () => Object.entries(CONFIG.DND5E.skills).map(i => ({label:
 const diceSizeOptions = [4, 6, 8, 10, 12, 20].map(i => ({label: 'd' + i, value: 'd' + i}));
 const teleportOptions = () => Object.entries(teleportEffects).map(i => ({label: i[1].name, value: i[0]}));
 const itemProperties = () => Object.entries(CONFIG.DND5E.itemProperties).map(i => ({label: i[1].label, value: i[0]}));
+const itemOptions = () => Object.entries(Item.implementation.compendiumBrowserTypes().physical.children).map(i => ({label: i[1].label, value: i[0]}));
 const armorOptions = () => Object.entries(CONFIG.DND5E.armorTypes).map(i => ({label: i[1], value: i[0]}));
 const spellSchoolOptions = () => Object.entries(CONFIG.DND5E.spellSchools).map(i => ({label: i[1].label, value: i[0]}));
 let baseWeaponOptions = [];
@@ -716,6 +717,7 @@ export let constants = {
     skillOptions,
     teleportOptions,
     itemProperties,
+    itemOptions,
     armorOptions,
     spellSchoolOptions,
     autoFailSaveEffectData,