+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 ## New Modern Content:
 - Ingenious Movement
 - Superior Atlas
+- Tinker's Magic
 ## Bug Fixes:
 - Generic Actions
 - Adventurer's Atlas
 - Guided Precision
 - Mapping Magic
+## API Update Notes
+- Added a helper for prompting a selection from the DND5e compendium browser.
 
 # 1.5.16 Release:
 ## New Legacy Content: